+++ frontend/catalyst/compilation_pipelines.py
@@ -59,9 +59,9 @@ from catalyst.utils.patching import Patcher
 # pylint: disable=unnecessary-lambda
 setattr(jax.interpreters.partial_eval.DynamicJaxprTracer, "__hash__", lambda x: id(x))
 
+# This flag cannot be set in ``QJIT.get_mlir()`` because values created before
+# that function is called must be consistent with the JAX configuration value.
 jax.config.update("jax_enable_x64", True)
-jax.config.update("jax_platform_name", "cpu")
-jax.config.update("jax_dynamic_shapes", True)
 
 
 def are_params_annotated(f: typing.Callable):